+++ pairtools_parquet/cli/dedup.py
@@ -8,12 +8,12 @@ from . import cli, common_io_options
 
 
 @cli.command()
-@click.argument("pairs_path", type=str, required=True)
+@click.argument("pairs_path", type=str, required=False)
 @click.option(
     "-o",
     "--output",
     type=str,
-    required=True,
+    default="",
     help="output file for pairs after duplicate removal."
     " The format follows the extension: .pairs, .pairs.gz or .parquet.",
 )