+++ pairtools_parquet/cli/flip.py
@@ -6,7 +6,7 @@ from . import cli, common_io_options
 
 
 @cli.command()
-@click.argument("pairs_path", type=str, required=True)
+@click.argument("pairs_path", type=str, required=False)
 @click.option(
     "-c",
     "--chroms-path",
@@ -21,7 +21,7 @@ from . import cli, common_io_options
     "-o",
     "--output",
     type=str,
-    required=True,
+    default="",
     help="output file. The format follows the extension: .pairs, .pairs.gz or .parquet.",
 )
 @click.option(