+++ pairtools_parquet/cli/sample.py
@@ -7,12 +7,12 @@ from . import cli, common_io_options
 
 @cli.command()
 @click.argument("fraction", type=float, required=True)
-@click.argument("pairs_path", type=str, required=True)
+@click.argument("pairs_path", type=str, required=False)
 @click.option(
     "-o",
     "--output",
     type=str,
-    required=True,
+    default="",
     help="output file. The format follows the extension: .pairs, .pairs.gz or .parquet.",
 )
 @click.option(